pyproject.toml
[project]
name = "psr-templates"
version = "0.1.0"
description = "PSR Template Test Harness"
requires-python = ">=3.8"
dependencies = [
    "tomli; python_version < '3.11'",
]

[project.scripts]
psr-build-template-structure = "arranger.run:main"
psr-prepare = "psr_prepare.cli:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages]
find = {where = ["src"]}

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.package-data]
arranger = ["templates/**/*"]

[dependency-groups]
dev = ["pytest", "pytest-mock", "pytest-cov", "flake8", "flake8-pyproject", "black", "python-semantic-release>=10.5.3", "mypy", "pre-commit"]

[tool.black]
line-length = 120
target-version = ['py38']

[tool.flake8]
max-line-length = 120
exclude = ["build", "dist", ".git", "__pycache__", ".pytest_cache", ".venv"]
ignore = ["E203", "W503"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
exclude = ["tests/", "tools/", "build/", "dist/"]

[[tool.mypy.overrides]]
module = ["tomllib", "importlib.resources", "jinja2"]
ignore_missing_imports = true
